refactor: move deck start view to feature
