+++ src/features/deck-start/index.ts
@@ -1,2 +1,3 @@
 export { DeckStartForm } from "./ui/DeckStartForm";
+export { DeckStartView } from "./ui/DeckStartView";
 export { useDeckFilterState } from "./model/useDeckFilterState";
