+++ src/features/deck-start/ui/DeckStartView.stories.tsx
@@ -1,11 +1,10 @@
-import { DeckStartForm } from "@/features/deck-start";
-
 import type { Meta, StoryObj } from "@storybook/react";
 import type { ComponentProps } from "react";
 
 import * as fixture from "@/storybook/fixture";
 import { INITIAL_VIEWPORTS } from "@/storybook/storybookViewports";
 
+import { DeckStartForm } from "./DeckStartForm";
 import { DeckStartView as View } from "./DeckStartView";
 
 type DeckStartFormProps = ComponentProps<typeof DeckStartForm>;
@@ -29,7 +28,7 @@ const deckStartForm: DeckStartFormProps = {
 };
 
 const meta = {
-  title: "Pages/Deck Start",
+  title: "Deck Start/DeckStartView",
   component: View,
   tags: ["autodocs"],
   parameters: { viewport: { viewports: INITIAL_VIEWPORTS, defaultViewport: "desktop" } },
